requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0,<2.0  # FastText requires NumPy <2.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Machine Learning & AI
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.0

# Deep Learning & Transformers
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
accelerate>=0.20.0
tensorboard>=2.12.0
fasttext-wheel>=0.9.2  # FastText for email classification

# Explainable AI (XAI)
lime>=0.2.0.1
shap>=0.41.0
eli5>=0.13.0

# Natural Language Processing
nltk>=3.7
textblob>=0.17.0
wordcloud>=1.8.0
fasttext-wheel>=0.9.2

# Web Framework (Dashboard için)
flask>=2.2.0
flask-cors>=3.0.10
dash>=2.6.0
plotly>=5.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Network & Security
requests>=2.28.0
ipaddress>=1.0.23
python-whois>=0.7.3

# Redis & Caching
redis>=4.5.0
hiredis>=2.2.0

# Utilities
tqdm>=4.64.0
colorama>=0.4.5
python-dotenv>=0.20.0

# Development & Testing
pytest>=7.1.0
pytest-cov>=4.0.0
jupyter>=1.0.0
ipython>=8.4.0

# Code Quality & Type Checking
flake8>=4.0.0
mypy>=0.990
black>=22.0.0
pylint>=2.13.0

# Data Processing
openpyxl>=3.0.10
xlsxwriter>=3.0.3
reportlab>=4.0.0

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.9.0
sqlmodel>=0.0.8

# REST API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-multipart>=0.0.6

# Optional but Recommended
psutil>=5.9.0
memory-profiler>=0.60.0